PythonService/pyproject.toml
[project]
name = "typeless-service"
version = "0.1.0"
description = "Typeless ASR and AI-powered text post-processing service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "google-generativeai>=0.8.0",
    "httpx>=0.28.1",
    "mlx-audio==0.3.0",
    "mlx-whisper>=0.4.3",
    "numpy<2.4",
    "openai>=1.0.0",
    "pydantic-settings>=2.0.0",
    "pydub>=0.25.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-multipart>=0.0.21",
    "slowapi>=0.1.9",
    "uvicorn[standard]>=0.40.0",
    "webrtcvad>=2.0.10",
    "websockets>=16.0",
]

[tool.uv.sources]
src = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.14",
]

[tool.pytest.ini_options]
pythonpath = "src"
testpaths = ["tests"]
